move adapters, serializers, and models into metrics folder
