+++ ui/app/models/metrics.js
@@ -1,0 +1,10 @@
+import { alias } from '@ember/object/computed';
+import { computed } from '@ember/object';
+import DS from 'ember-data';
+import lazyCapabilities, { apiPath } from 'vault/macros/lazy-capabilities';
+
+let { attr } = DS;
+
+export default DS.Model.extend({
+  // console.log("getting hit")
+});
