+++ ui/app/models/metrics/http-requests.js
